feat: Add Jest setup file and update test environment
chore: Update package.json with Jest canvas and jsdom dependencies

feat: Add canvas module for setting up game canvas

feat: Add player module with player interface and image assets

test: Add unit test for updating slide number

feat: Add event handlers for player movement and slide navigation

test: Remove unnecessary test for main function

feat: Add drawing module with image caching and drawing functions
