+++ jest.setup.js
@@ -1,0 +1,1 @@
+global.Image = window.Image;
