+++ package.json
@@ -16,6 +16,8 @@
     "eslint-config-prettier": "~9.1",
     "eslint-plugin-jest": "~28.8.2",
     "jest": "~29.7",
+    "jest-canvas-mock": "^2.5.2",
+    "jest-environment-jsdom": "^29.7.0",
     "nodemon": "^3.1.4",
     "prettier": "~3.3.3",
     "rimraf": "~5.0",
