+++ src/main.ts
@@ -1,4 +1,3 @@
 import { gameLoop } from './drawBilboard';
 
-console.log('Hello, world! 2');
 gameLoop();
